(PC-7937) remove useless code
